geotiff: extract read_vrt and VRT dask helpers to _backends/vrt.py (#1887)
Step 9 of #1813's multi-PR refactor of __init__.py. Pure code motion;
no public API change.

Moved into a new xrspatial/geotiff/_backends/vrt.py:

- read_vrt: VRT read entry point (~250 lines), with the eager-path
  CRS/transform/holes attrs population, the per-band nodata
  selection (#1598), and the integer-sentinel post-decode mask
  (#1611, #1825) intact.
- _vrt_chunk_read: per-chunk delayed reader used by the dask graph.
- _read_vrt_chunked: lazy dask dispatch when chunks= is set (#1814),
  including the parse-time hole detection and 50k-task graph cap.
- _MAX_VRT_DASK_CHUNKS constant.

The XML parsing, source-path containment, per-source decode, and
``_apply_integer_sentinel_mask`` already live in
xrspatial/geotiff/_vrt.py (untouched here); this module orchestrates
only the eager / dask / GPU dispatch around them.

Function bodies are unchanged except for adjusted relative imports
(``._X`` -> ``.._X``). __init__.py re-exports ``read_vrt`` so the
public surface is unchanged.

Knock-on cleanup
================

``_backends/dask.py`` had a lazy ``from .. import read_vrt`` (added in
#1886) because ``read_vrt`` was still in __init__.py. With ``read_vrt``
now in a sibling backend module, the import is promoted to a static
top-of-module ``from .vrt import read_vrt`` and the dispatch site
becomes a direct call.

Diff stats
==========

- __init__.py: 2505 -> 1905 lines (-600).
- New _backends/vrt.py: 624 lines.

Verification: pixel-parity matrix from #1889 (192 cells), runtime
identity (9), VRT-specific tests (lazy chunks #1814, lazy chunks
#1798, window validation #1697, backend coverage), full geotiff
suite minus the 8 pre-existing main failures -- 2862/2862 pass.

Refs #1813.
